+++ services/horizon/CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this
 file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Changes
+* Return inner and outer result codes for fee bump transactions ([4081](https://github.com/stellar/go/pull/4081))
+
 ## v2.11.0
 
 ### Changes